requirements.txt
pydantic-ai>=1.62.0
python-telegram-bot>=21.0
httpx>=0.27.0
trafilatura>=2.0.0
playwright>=1.49.0
python-dateutil>=2.9.0
qdrant-client>=1.12.0
sentence-transformers>=3.3.0
